requirements.txt
anthropic>=0.34.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.104.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
tenacity>=8.2.3
